requirements.txt

### requirements.txt

```plaintext
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==3.14.0
